feat: move command bar helpers
