+++ app/javascript/dashboard/composables/commands/useGoToCommandHotKeys.js
@@ -19,7 +19,7 @@ import {
   ICON_TEAM_REPORTS,
   ICON_USER_PROFILE,
   ICON_CONVERSATION_REPORTS,
-} from './helpers/icons';
+} from 'dashboard/helper/CommandBar/icons';
 import { frontendURL } from 'dashboard/helper/URLHelper';
 import { FEATURE_FLAGS } from 'dashboard/featureFlags';
