+++ app/javascript/dashboard/helper/CommandBar/actions.js
