+++ app/javascript/dashboard/helper/CommandBar/events.js
