+++ app/javascript/dashboard/helper/CommandBar/icons.js
